+++ SubtitleRenderer/pngrenderer.hpp
@@ -15,7 +15,7 @@ class PNGRenderer
 {
 public:
     PNGRenderer();
-    PNGRenderer(const std::string &text, const std::string &fontFamily = {}, int fontSize = 48, int furiganaFontSize = 20);
+    PNGRenderer(const std::string &text, const std::string &fontFamily = {}, unsigned long fontSize = 48, unsigned long furiganaFontSize = 20);
     ~PNGRenderer() = default;
 
     enum class TextJustify
@@ -65,7 +65,7 @@ public:
         _fontFamily = fontFamily;
     }
 
-    inline void setFontSize(int fontSize)
+    inline void setFontSize(unsigned long fontSize)
     {
         _fontSize = fontSize;
     }
@@ -75,7 +75,7 @@ public:
         _fontColor = fontColor;
     }
 
-    inline void setFuriganaFontSize(int furiganaFontSize)
+    inline void setFuriganaFontSize(unsigned long furiganaFontSize)
     {
         _furiganaFontSize = furiganaFontSize;
     }
@@ -146,12 +146,12 @@ public:
         _borderColor = borderColor;
     }
 
-    inline void setBorderSize(int borderSize)
+    inline void setBorderSize(unsigned long borderSize)
     {
         _borderSize = borderSize;
     }
 
-    inline void setFuriganaBorderSize(int furiganaBorderSize)
+    inline void setFuriganaBorderSize(unsigned long furiganaBorderSize)
     {
         _furiganaBorderSize = furiganaBorderSize;
     }
@@ -162,17 +162,17 @@ private:
     bool _vertical = false;
     std::string _text;
     std::string _fontFamily;
-    int _fontSize = 48;
+    unsigned long _fontSize = 48;
     std::string _fontColor = "#f1f1f1";
-    int _furiganaFontSize = 20;
+    unsigned long _furiganaFontSize = 20;
     std::string _furiganaFontColor = "#f1f1f1";
     int _lineSpaceReduction = 0;
     int _furiganaLineSpaceReduction = 0;
     TextJustify _textJustify = TextJustify::Center;
     FuriganaDistance _furiganaDistance = FuriganaDistance::Unchanged;
     std::string _borderColor = "#191919";
-    int _borderSize = 4;
-    int _furiganaBorderSize = 2;
+    unsigned long _borderSize = 4;
+    unsigned long _furiganaBorderSize = 2;
 };
 
 #endif // PNGRENDERER_HPP
